Update middleCS490Beta.php
Updated and sends back and forth. There are three final issues occurring before this is fully completed. These final issues are just array to string conversion errors to the back and Uninitialized string offset: 0 all on line 174 occurring in the back. I'm going to be looking further into this and will see what I can do to complete this issue. I also shifted the exec to send the results into a value then exploding that value, which provided me with the ability to read each value within the $resultSet so it would then get values to check.