Update middleCS490Beta.php
Leaving a further update that I found the final issue. It's centered around a reoccurring issue on the usage of file_put_contents not working. In cases when it's on a local system, it works fine but when it's moved then it causes issues. I'm researching how to get around this issue (Might totally just use fopen/ fclose and add the information in that way) and once it's done it should allow the program to be executed. Once that occurs the rest should check with one another and we should have a completed project.